+++ Sources/LiveKit/Core/Room+AppStateDelegate.swift
@@ -1,0 +1,61 @@
+/*
+ * Copyright 2024 LiveKit
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import Foundation
+
+extension Room: AppStateDelegate {
+    func appDidEnterBackground() {
+        guard _state.roomOptions.suspendLocalVideoTracksInBackground else { return }
+
+        let cameraVideoTracks = localParticipant.localVideoTracks.filter { $0.source == .camera }
+
+        guard !cameraVideoTracks.isEmpty else { return }
+
+        Task.detached {
+            for cameraVideoTrack in cameraVideoTracks {
+                do {
+                    try await cameraVideoTrack.suspend()
+                } catch {
+                    self.log("Failed to suspend video track with error: \(error)")
+                }
+            }
+        }
+    }
+
+    func appWillEnterForeground() {
+        let cameraVideoTracks = localParticipant.localVideoTracks.filter { $0.source == .camera }
+
+        guard !cameraVideoTracks.isEmpty else { return }
+
+        Task.detached {
+            for cameraVideoTrack in cameraVideoTracks {
+                do {
+                    try await cameraVideoTrack.resume()
+                } catch {
+                    self.log("Failed to resumed video track with error: \(error)")
+                }
+            }
+        }
+    }
+
+    func appWillTerminate() {
+        // attempt to disconnect if already connected.
+        // this is not guranteed since there is no reliable way to detect app termination.
+        Task.detached {
+            await self.disconnect()
+        }
+    }
+}
